% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKinshipMatrix.R
\name{createKinshipMatrix}
\alias{createKinshipMatrix}
\title{createKinshipMatrix: create a kinship matrix using ibs in R}
\usage{
createKinshipMatrix(gwaa.data = NULL, plot.mds = T)
}
\arguments{
\item{gwaa.data}{GenABEL gwaa.data object.}

\item{plot.mds}{Default TRUE. Creates a plot of classical multidimensional
scaling (MDS) of the kinship matrix.}
}
\description{
createKinshipMatrix: create a kinship matrix using ibs in R
}
