% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumuPos.R
\name{CumuPos}
\alias{CumuPos}
\title{CumuPos: Determine the cumulative positions of scan.gwaa.object for plotting.}
\usage{
CumuPos(scan.gwaa.object)
}
\arguments{
\item{scan.gwaa.object}{GenABEL scan.gwaa.object object containing GWAS results.}
}
\description{
CumuPos: Determine the cumulative positions of scan.gwaa.object for plotting.
}
