% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAdjacentLD.R
\name{estimateLD}
\alias{estimateLD}
\title{estimateLD: Estimated LD (r2) between adjacent markers.}
\usage{
estimateLD(markerfile, snpvec = NULL, idvec = NULL, full.results = F,
  ld.window.kb = 1000, outfile = NULL)
}
\arguments{
\item{markerfile}{Path to markerfile}

\item{snpvec}{Vector of SNPs to include (all others excluded). Optional}

\item{idvec}{Vector of IDs to include (all others excluded). Optional}

\item{full.results}{Default FALSE. If TRUE, prints the table of LD measures
between adjacent loci.}

\item{ld.window.kb}{Default 1000. Maximum kb distance between which to
measure LD. Higher values will ensure all loci are captured, but will take
longer.}
}
\description{
estimateLD: Estimated LD (r2) between adjacent markers.
}
