% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQTLEffects.R
\name{estimateQTLEffects}
\alias{estimateQTLEffects}
\title{estimateQTLEffects: estimate variance attributed to simulated QTLs}
\usage{
estimateQTLEffects(markerfile, phenofile, flanking.window, merged = F)
}
\arguments{
\item{markerfile}{string}

\item{phenofile}{string}

\item{flanking.window}{integer. Number of SNPs flanking the QTL for
estimation of effect sizes}

\item{merged}{Boolean, FALSE. If TRUE, include the quantitative trait
nucleotide from the simulation.}
}
\description{
estimateQTLEffects: estimate variance attributed to simulated QTLs
}
