% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGwasSummary.R
\name{FullGwasSummary}
\alias{FullGwasSummary}
\title{FullSummary: Create a Manhatten and P-P plots and return top results from a
GenABEL scan.gwaa.object.}
\usage{
FullGwasSummary(scan.gwaa.object, corrected = FALSE, bonf = F,
  qtl.ids = NULL)
}
\arguments{
\item{scan.gwaa.object}{GenABEL scan.gwaa.object object containing GWAS
results.}

\item{corrected}{logical. If true, GWAS results will be reported after
correction for genomic control.}

\item{bonf}{numeric. Threshold P-value for genome-wide significance. If not
defined, will be calculated using a Bonferroni correction.}

\item{qtl.ids}{vector. Vector of SNP names for QTL loci (if they are included
in the dataset). If specified, their positions will be highlighted.}
}
\description{
FullSummary: Create a Manhatten and P-P plots and return top results from a
GenABEL scan.gwaa.object.
}
