% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullPpPlot.R
\name{FullPpPlot}
\alias{FullPpPlot}
\title{FullPpPlot: Create a P-P plot from a GenABEL scan.gwaa.object.}
\usage{
FullPpPlot(scan.gwaa.object, corrected = FALSE, lambda = NULL,
  qtl.ids = NULL)
}
\arguments{
\item{scan.gwaa.object}{GenABEL scan.gwaa.object object containing GWAS
results.}

\item{corrected}{logical. If true, GWAS results will be reported after
correction for genomic control.}

\item{lambda}{numeric. Genomic control parameter, to be specified if a
different value from the scan.gwaa.object is to be used.}

\item{qtl.ids}{vector. Vector of SNP names for QTL loci (if they are included
in the dataset). If specified, their positions will be highlighted.}
}
\description{
FullPpPlot: Create a P-P plot from a GenABEL scan.gwaa.object.
}
