% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot: Multiple plot function. # ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
multiplot: Multiple plot function. # ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
