% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseParamFile.R
\name{parseParamFile}
\alias{parseParamFile}
\title{parseParamFile: Parse the parameter file.}
\usage{
parseParamFile(paramfile, markerfile = NULL)
}
\arguments{
\item{paramfile}{Path to the parameter file}

\item{markerfile}{If provided, then will assume param file is in the same
directory and parse the *.prm file.}
}
\description{
parseParamFile: Parse the parameter file.
}
