% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGCTA.R
\name{RunGCTA}
\alias{RunGCTA}
\title{RunGCTA: Run GCTA}
\usage{
RunGCTA(input.string)
}
\arguments{
\item{input.string}{Command to be passed to GCTA. Do not include "gcta64" at
the beginning of the beginning of the string e.g. RunGCTA("--help")}
}
\description{
RunGCTA: Run GCTA
}
