% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPLINK.R
\name{RunPLINK}
\alias{RunPLINK}
\title{RunPLINK: Run PLINK}
\usage{
RunPLINK(input.string)
}
\arguments{
\item{input.string}{Command to be passed to PLINK. Do not include "plink" at
the beginning of the beginning of the string e.g. RunPLINK("--help")}
}
\description{
RunPLINK: Run PLINK
}
