% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQMSim.R
\name{RunQMSim}
\alias{RunQMSim}
\title{RunQMSim: Run QMSim}
\usage{
RunQMSim(input.string)
}
\arguments{
\item{input.string}{Command to be passed to QMSim Do not include "gcta64" at
the beginning of the beginning of the string e.g. RunGCTA("--help")}
}
\description{
RunQMSim: Run QMSim
}
