% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGenomicHeritability.R
\name{estimateGenomicHeritability}
\alias{estimateGenomicHeritability}
\title{estimateGenomicHeritability: create a kinship matrix using ibs in R}
\usage{
estimateGenomicHeritability(markerfile, merged = F)
}
\arguments{
\item{merged}{Default FALSE. If merged, will include the QTL markers.}
}
\description{
estimateGenomicHeritability: create a kinship matrix using ibs in R
}
