% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAdjacentLD.R
\name{estimateLD}
\alias{estimateLD}
\title{estimateAdjacentLD: Estimated LD (r2) between adjacent markers.}
\usage{
estimateLD(markerfile, adjacent.only = T, full.results = F,
  ld.window.kb = 1000)
}
\arguments{
\item{full.results}{Default FALSE. If TRUE, prints the table of LD measures
between adjacent loci.}

\item{ld.window.kb}{Default 1000. Maximum kb distance between which to
measure LD. Higher values will ensure all loci are captured, but will take
longer.}
}
\description{
estimateAdjacentLD: Estimated LD (r2) between adjacent markers.
}
