% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseQMSIM.R
\name{parseQMSIM}
\alias{parseQMSIM}
\title{parseQMSIM: Parse files in a QMSIM directory.}
\usage{
parseQMSIM(model.name, markerfile = "p1_mrk_001.txt",
  phenofile = "p1_data_001.txt", flanking.window = 10)
}
\arguments{
\item{model.name}{string. This is the name of the directory containing the simulation information.}

\item{markerfile}{string. Default is "p1_mrk_001.txt".}

\item{phenofile}{string. Default is "p1_data_001.txt"}

\item{flanking.window}{integer}
}
\description{
parseQMSIM: Parse files in a QMSIM directory.
}
